requirements.txt
apify-client==1.8.1
openai==1.51.1
langchain==0.3.1
langchain_community==0.3.1
langchain_openai==0.2.2
tiktoken==0.8.0
gradio==4.44.1
python-docx
docx2txt==0.8
beautifulsoup4
markdown
httpx==0.27.2










